ATC-138: split cli.ts into cli/ command modules; root assembly moves to main.ts
