+++ app-server/src/main.ts
@@ -2,11 +2,33 @@ import { BunRuntime, BunServices } from "@effect/platform-bun"
 import { Effect, Layer } from "effect"
 import { Command } from "effect/unstable/cli"
 import * as BuildInfo from "./buildInfo.ts"
-import { atc } from "./cli.ts"
+import { fs, health, version } from "./cli/cli.ts"
+import { api, capabilities, context } from "./cli/gateway.ts"
+import { project } from "./cli/projects.ts"
+import { serve } from "./cli/serve.ts"
+import { terminal } from "./cli/terminals.ts"
+import { smoke } from "./smoke.ts"
 import * as Subprocess from "./subprocess.ts"
 
-// The only runMain in the application. Everything else is Layers and Effects,
-// so even the CLI version resolves through the injected BuildInfo service.
+// The entrypoint: root command assembly plus the only runMain in the
+// application. Everything else is Layers and Effects, so even the CLI
+// version resolves through the injected BuildInfo service.
+const atc = Command.make("atc").pipe(
+  Command.withDescription("ATC App Server"),
+  Command.withSubcommands([
+    serve,
+    api,
+    context,
+    capabilities,
+    health,
+    version,
+    project,
+    terminal,
+    fs,
+    smoke,
+  ]),
+)
+
 Effect.gen(function* () {
   const build = yield* BuildInfo.BuildInfo
   yield* Command.run(atc, { version: build.version })
